requirements.txt
openai==2.29.0
python-dotenv==1.2.2
streamlit==1.55.0
